fix: tighten quoted zap validation
Lock deposit and increase flows to quoted mode and price anchors, derive withdraw swap thresholds from the quoted price, and align tests with the new quoted zap semantics.

Made-with: Cursor

+++ programs/lp_handler/src/state/events.rs
@@ -108,7 +108,7 @@ pub struct SecurityConfigClosed {
     /// 接收租金的地址
     pub receiver: Pubkey,
     /// 关闭前的 pool 数量
-    pub pools_count: usize,
+    pub pools_count: u32,
     /// 关闭前的 fee_owners 数量
-    pub fee_owners_count: usize,
+    pub fee_owners_count: u32,
 }
